requirements.txt
sanic
spitfire3
requests
tortoise-orm
sib-api-v3-sdk
python-dotenv